requirements.txt
pytest==8.3.3
selenium==4.25.0
webdriver-manager==4.0.2
requests==2.32.3
allure-pytest==2.13.5